+++ src/test/java/ee/sk/smartid/SmartIdClientTest.java
@@ -358,6 +358,7 @@ class SmartIdClientTest {
                     .initAuthenticationSession();
 
             URI qrCodeUri = new DeviceLinkBuilder()
+                    .withSchemeName("smart-id-demo")
                     .withDeviceLinkBase(response.getDeviceLinkBase().toString())
                     .withDeviceLinkType(deviceLinkType)
                     .withSessionType(SessionType.AUTHENTICATION)
@@ -385,6 +386,7 @@ class SmartIdClientTest {
             long elapsedSeconds = Duration.between(response.getReceivedAt(), Instant.now()).getSeconds();
 
             URI qrCodeUri = new DeviceLinkBuilder()
+                    .withSchemeName("smart-id-demo")
                     .withDeviceLinkBase(response.getDeviceLinkBase().toString())
                     .withDeviceLinkType(DeviceLinkType.QR_CODE)
                     .withSessionType(SessionType.AUTHENTICATION)